create fams dash via dazzler builder
